requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.5
PyYAML==6.0.2
APScheduler==3.10.4
pydantic==2.10.4
python-multipart==0.0.20
requests==2.32.3
yfinance==0.2.54
python-dotenv==1.0.1
